UX-469 🐛 - fixing - failing tests  / 🦄.
